file structure rename and reorder. updated .gitignore.
