+++ .gitignore
@@ -310,6 +310,4 @@ pyvenv.cfg
 
 # End of https://www.gitignore.io/api/c,vim,c++,linux,python,windows,pycharm,notepadpp,powershell
 
-testing/test.txt # test output files
-testing/test.json# test output files
-/original/~$*.xlsx
+data_analytics/*
