+++ src/libs/hal.py
